services/product/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pymongo==3.12.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pydantic==1.8.2
python-dotenv==0.19.0
pytest==6.2.5
httpx==0.19.0 